+++ lib/graphql/schema/member/has_arguments.rb
@@ -103,20 +103,24 @@ module GraphQL
             context.schema.object_from_id(id, context)
           end
 
-          def load_application_object(argument, lookup_as_type, id)
+          def load_application_object(arg_kwarg, id)
+            argument = @arguments_by_keyword[arg_kwarg]
+            lookup_as_type = @arguments_loads_as_type[arg_kwarg]
             # See if any object can be found for this ID
             loaded_application_object = object_from_id(lookup_as_type, id, context)
             context.schema.after_lazy(loaded_application_object) do |application_object|
-              begin
-                if application_object.nil?
-                  raise LoadApplicationObjectFailedError.new(argument: argument, id: id, object: application_object)
-                end
-                # Double-check that the located object is actually of this type
-                # (Don't want to allow arbitrary access to objects this way)
-                application_object_type = context.schema.resolve_type(lookup_as_type, application_object, context)
+              if application_object.nil?
+                err = LoadApplicationObjectFailedError.new(argument: argument, id: id, object: application_object)
+                load_application_object_failed(err)
+              end
+              # Double-check that the located object is actually of this type
+              # (Don't want to allow arbitrary access to objects this way)
+              resolved_application_object_type = context.schema.resolve_type(lookup_as_type, application_object, context)
+              context.schema.after_lazy(resolved_application_object_type) do |application_object_type|
                 possible_object_types = context.schema.possible_types(lookup_as_type)
                 if !possible_object_types.include?(application_object_type)
-                  raise LoadApplicationObjectFailedError.new(argument: argument, id: id, object: application_object)
+                  err = LoadApplicationObjectFailedError.new(argument: argument, id: id, object: application_object)
+                  load_application_object_failed(err)
                 else
                   # This object was loaded successfully
                   # and resolved to the right type,
@@ -137,9 +141,6 @@ module GraphQL
                     application_object
                   end
                 end
-              rescue LoadApplicationObjectFailedError => err
-                # pass it to a handler
-                load_application_object_failed(err)
               end
             end
           end