+++ lib/graphql/types/relay/nodes_field.rb
@@ -4,7 +4,23 @@ module GraphQL
     module Relay
       # This can be used for implementing `Query.nodes(ids: ...)`,
       # or use it for inspiration for your own field definition.
-      # @see GraphQL::Types::Relay::NodeField
+      #
+      # @example Adding this field directly
+      #   add_field(GraphQL::Types::Relay::NodesField)
+      #
+      # @example Implementing a similar field in your own Query root
+      #
+      #   field :nodes, [GraphQL::Types::Relay::Node, null: true], null: false,
+      #     description: Fetches a list of objects given a list of IDs." do
+      #       argument :ids, [ID], required: true
+      #     end
+      #
+      #   def nodes(ids:)
+      #     ids.map do |id|
+      #       context.schema.object_from_id(context, id)
+      #     end
+      #   end
+      #
       NodesField = GraphQL::Schema::Field.new(
         name: "nodes",
         owner: nil,